feat: Refactor blob storage handling and update related models
- Updated connection string configurations in launch.json for blob storage.
- Removed the ContainerNames class as it is no longer needed.
- Refactored IBlobBus interface methods to use path instead of container names for better clarity.
- Enhanced BlobBusBase implementation to align with the new interface structure.
- Updated AzureStorageBlobBus to handle new response types and improved error logging.
- Added LocalStorageBlobBus registration in BlobModule for local storage handling.
- Adjusted various request and response models to accommodate changes in blob file handling.
- Improved documentation and XML comments for clarity on new implementations.

+++ Persistence/PersistenceDb.Models/Administration/Region.cs
@@ -15,7 +15,7 @@ public class Region
     /// Id de Región
     /// </summary>
     /// <value></value>
-    public int Id { get; set; }
+    public short Id { get; set; }
 
     [Required]
     [StringLength(20)]
@@ -42,7 +42,7 @@ public class Region
     /// Id de País
     /// </summary>
     /// <value></value>
-    public int CountryId { get; set; }
+    public byte CountryId { get; set; }
 
     [Required]
     [Column("REG_ESTADO")]
